Fix the docker-compose serving (#694)
This fixes a grab bag of problems for me, but still doesn't get it to
usability. The static files makes this render.

And then the other changes address the errors in logs

```
dab_postgres  | 2025-02-05 18:39:53.062 UTC [58] FATAL:  role "$${POSTGRES_USER}" does not exist
```

and

```
test_app-1    | uwsgi_check_logrotate()/lseek(): Illegal seek [core/logging.c line 494]
```

This is addressed by the uwsgi config file change, and to the startup
script.

I think this is some progress, but unfortunately, I'm still blocked on
using the docker-compose environment as I get a CSRF error when trying
to log in.