AAP-37017 Fix how we build AuthExceptions (#664)
The `AuthException` class from `social_core.exceptions` takes a first
parameter of a backend and a second of a message. Since we only had one
parameter when we constructed our AuthException, printing it may be
black:

```
In [6]: from social_core.exceptions import AuthException
In [7]: try:
   ...:     raise AuthException(None, 'Testing')
   ...: except AuthException as e:
   ...:     print(f'AuthException: {e}')
   ...: 
AuthException: Testing

In [8]: try:
   ...:     raise AuthException('Testing')
   ...: except AuthException as e:
   ...:     print(f'AuthException: {e}')
   ...: 
AuthException: 
```

This PR fixes this by passing in `None` as the first backend parameter
to the `AuthException`.