[AAP-47730] Make SonarQube relocatable (#772)
## Description
Following the advice of
https://github.com/ansible-automation-platform/cicd-ops/blob/main/doc/sonar.md#sonar-projectproperties,
compute the values of the SonarQube projectKey and organization
variables at run time.

Addresses https://issues.redhat.com/browse/AAP-47730

## Type of Change
<!-- Mandatory: Check one or more boxes that apply -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update
- [X] Test update
- [ ] Refactoring (no functional changes)
- [ ] Development environment change
- [X] Configuration change

## Self-Review Checklist
<!-- These items help ensure quality - they complement our automated CI
checks -->
- [X] I have performed a self-review of my code
- [X] I have added relevant comments to complex code sections
- [ ] I have updated documentation where needed
- [X] I have considered the security impact of these changes
- [ ] I have considered performance implications
- [ ] I have thought about error handling and edge cases
- [X] I have tested the changes in my local environment

### Steps to Test
The very nature of this PR is a test; you'd need to commit it or
something like it in order to test. In DAB's case, the affected workflow
only runs on a merge, and it's a separate action outside the purview of
a PR merge (i.e. even if this fails, it will not affect the ability to
merge subsequent PRs).

### Expected Results
SonarQube scan results should appear in whatever organization you are
running in - i.e. ansible vs. ansible-automation-platform

<!-- ### Screenshots/Logs -->
<!-- Add if relevant to demonstrate the changes -->

---------

Signed-off-by: David Schmidt <51931019+schmidtd@users.noreply.github.com>