AAP-50626 AAP-50627 Use ansible_id references in access lists (#783)
This is needed to complete the UI connection of the access lists.

Background:

Due to technical issues, we have asked the UI to use the component-based
access list for resources owned by a certain component. So if there's an
AWX resource, get the access list served from AWX for that item, like an
inventory or job template.

However, our endpoints were still using the local team or user id. This
is a problem, because all other RBAC management is centralized around
the resource server. So a translation is needed, and right now there's
no easy way to do that translation. So the solution (this) is to
formulate all the access list related endpoints around the ansible_id of
the actor, that being the user or the team.

Corresponding to this, we need a fix for the bug
https://github.com/ansible/django-ansible-base/issues/534, so that the
client has the ansible_id as a starting point when navigating to the
component endpoints

---------

Co-authored-by: John Westcott IV <32551173+john-westcott-iv@users.noreply.github.com>