AAP-50614 Add a methods for user claims serialization and hashing (#787)
## Description
Up until this point, the DAB app jwt_consumer has housed logic for
_saving_ user claims to the RBAC models in the local database.

We have planned a change which will modify this contract. A service will
still get serialized claims from the resource server and save them by
existing jwt_consumer logic, but when this happens we will store a hash
of those claims (producible from the claims data, no DB interaction
needed) in the local cache. Then, we will cross-check against this hash
for future requests and do nothing if it matches. This will allow us to
move the transmission of the cache data to a new mechanism which avoids
the header size limitations that plague JWT usage right now.

This just does this first part - move _generation_ of claims data into
DAB. Before this we _consumed_ the claims data, and this remains
unchanged. Additionally, we add some simple logic to _hash_ the claims
data and test that the hash is consistent.