[AAP-50642] memoize pem private key to avoid openssl slowdown (#791)
For 2.6, we've moved to RHEL9 for the base container, which has openssl
3.2 in it. Apparently , its is a known issue that the RSA key validation
is more secure and more slow in this version vs. Openssl 1 that RHEL8
uses. We were loading and validating the same jwt private key on every
request, at a cost of ~ 38 ms, vs previously this took ~ 3 ms. We
noticed this slowdown in the benchmarks and after deep dive/profiling we
found this was the difference and memoizing the validated private key
saves us from this broad slowdown. See
https://issues.redhat.com/browse/AAP-50642 for more details

[pyca/cryptography#7236
(comment)](https://github.com/pyca/cryptography/issues/7236#issuecomment-1131908472)
<--- report that details Openssl 3 slowdown