AAP-51654 Enforce that references to objects are strictly done in claims data, not token data (#809)
The prior solution partially migrate us from passing claims data inside
of the JWT data to a new user "claims" data structure.

However, it was still referencing "objects" inside of the JWT data
itself.

This refactors the methods out of the auth class, which makes this kind
of bug likely. Stuff is just passed as method arguments in this new
approach. This is what I originally expected out of this work.