AAP-51883 Fix unsafe assumption about org content type (#814)
This fixes a server error

This is due to, specifically, the registration of the `Team` model in
galaxy_ng with no parent model.


https://github.com/ansible/galaxy_ng/blob/5fce949deb374e1755029f7d9a5a6d109d031e69/galaxy_ng/app/models/__init__.py#L50-L59

Knowing that the parent model is null, you can read the traceback in a
way that makes sense. The variable `team_parent_model` gets a value of
`None`, which is 100% unexpected.

At the end, `model` was _intended_ to be passed, but it has a value of
`None`, which the method interprets as not being provided.