AAP-54064 Delete DABPermission entries if model is not registered and not used (#851)
This is specifically needed for a prior behavior of `Team` model
auto-registration. This was deleted a few months ago, but old versions
may still create permissions for the team model.

Re-stating, that leaves a scenario where an app might _enable RBAC_ but
_register no models_ and then fail to upgrade. This is written
specifically to address that situation. Previously, we just threw an
error out of an abundance of caution (which was correct, because we
didn't have a specific action clarified).

For this specific case, the remediation action is clear - delete the
permission entries. We still have to be really specific to avoid
deleting unintended things. The criteria should be well-formed in this
patch.