AAP-53278 General API optimizations for RBAC (#837)
This does general, obligatory, optimizations of API endpoints that were
recently added.

The largest code change is because using `LARGE=true` with the bootstrap
script wasn't creating all the RBAC objects, additional logic to do so
was added. This did help to test cases of >25 Role definitions,
role-user-assignments, and role-team-assignments.

This work is basically never-ending, so any stopping point is arbitrary.
However, the new `/service-index/` views greatly benefit from this.

The found solution often repeated for `resource__content_type` also
suggests that the resource summary_fields entry is _generally_ not
well-optimized, because it's unlikely someone would have continued that
to include `__content_type`.